requirements.txt
starcli
git+https://github.com/FortuneDayssss/BilibiliUploader.git@master#egg=BilibiliUploader
gitpython
docker
jupyter
wecover